requirements.txt
opencv-python-headless
scikit-image
fpdf
Pillow